robust random
